requirements.txt
Flask==2.3.2
pytest==7.4.0
blue==0.9.1
google-cloud-bigquery==3.11.4
numpy==1.24.3
pydub==0.25.1
resampy==0.4.2
tensorflow==2.13.0
tf-slim==1.1.0
six==1.16.0
pydantic==1.10.11
soundfile==0.12.1
audioflux==0.1.6
pandas==2.0.3
flatten_json==0.1.13
scikit_learn==1.3.0
xgboost==1.7.6
matplotlib==3.6.0
seaborn==0.12.2
db-dtypes==1.1.1 
tqdm==4.66.1
lightgbm==4.0.0
plotnine==0.12.3
wordcloud==1.9.2
shap==0.42.1
google_cloud_storage==2.11.0
ffmpeg-python==0.2.0
essentia==2.1b6.dev1034; sys_platform=="linux"
